Disable closing idle Ssh connections again
By switching to Apache SSHD 0.6.0 in
812a823569339ec6b839a5c599b9bac7be73d8c6, we started to close idle Ssh
connections after 10 minutes of inactivity (See r1163164 of Apache
SSHD). This closing of idle connections caused problems for scripts
that listened for streamed events. Even though the scripts
reconnected, and gerrit accepted the connection attempts, sometimes no
events got passed through the new connections and the task queue
started to fill up.

Until we can isolate the root cause of the problem, we revert to the
old behaviour of not closing idle connections, and thereby mitigate
the problem.

Additionally, we allow to configure the idle timeout.

Change-Id: Ia4f226771b082057ecb45e6c2a9a8d809ec1389d